// src/config/media-status.enum.ts
export enum MediaStatus {
	UPLOAD_IN_PROGRESS = 'UPLOAD_IN_PROGRESS',
	WAITING_TRANSCODING_START = 'WAITING_TRANSCODING_START',
	TRANSCODING_IN_PROGRESS = 'TRANSCODING_IN_PROGRESS',
	WAITING_FINALIZE_ABR_MEZZANINE_START = 'WAITING_FINALIZE_ABR_MEZZANINE_START',
	FINALIZE_ABR_MEZZANINE_IN_PROGRESS = 'FINALIZE_ABR_MEZZANINE_IN_PROGRESS',
	PUBLICATION_IN_PROGRESS = 'PUBLICATION_IN_PROGRESS',
	READY = 'READY',
	ERROR = 'ERROR',
	PUBLISHED = 'PUBLISHED'
}